starting final-project

+++ staff/joel-manzano/full-stack/demo-proyect/logic/src/deleteUser.js
@@ -1,9 +1,18 @@
-const { models: { User } } = require('data')
-const { validators: { validateId } } = require('commons')
+const { models: { User, Note } } = require('data')
+const { validators: { validateId, validatePassword } } = require('commons')
 
-function deleteUser(id) {
-    validateId(id)
-    return User.deleteOne({id})
-}
+
+function deleteUser(userId, password) {
+    validateId(userId)
+    validatePassword(password)
+
+    return Note.deleteMany({ user: userId })
+        .then(() => {
+            return User.deleteOne({ id: userId, password })
+            .then(result => {
+                if (result.deletedCount === 0) throw new Error(`wrong user ${userId} or password`)
+            })
+        })
+    }
 
 module.exports = deleteUser
